added user object to profile servlet

+++ src/main/java/com/codeup/adlister/dao/MySQLAdsDao.java
@@ -39,11 +39,11 @@ public class MySQLAdsDao implements Ads {
         }
     }
 
-    public List<Ad> userAll(int userId) { // used to display user ads on user profile
+    public List<Ad> userAll(long userId) { // used to display user ads on user profile
         PreparedStatement stmt = null;
         try {
             stmt = connection.prepareStatement("SELECT * FROM ads WHERE user_id = ?");
-            stmt.setInt(1, userId);
+            stmt.setLong(1, userId);
             ResultSet rs = stmt.executeQuery();
             return createAdsFromResults(rs);
         } catch (SQLException e) {
